+++ src/Fire.js
@@ -15,6 +15,8 @@
 
 const {Gio, GObject} = imports.gi;
 
+const _ = imports.gettext.domain('burn-my-windows').gettext;
+
 const ExtensionUtils = imports.misc.extensionUtils;
 const Me             = imports.misc.extensionUtils.getCurrentExtension();
 const utils          = Me.imports.src.utils;
@@ -53,7 +55,7 @@ var Fire = class Fire {
   // This will be shown in the sidebar of the preferences dialog as well as in the
   // drop-down menus where the user can choose the effect.
   static getLabel() {
-    return 'Fire';
+    return _('Fire');
   }
 
   // -------------------------------------------------------------------- API for prefs.js
@@ -125,7 +127,7 @@ var Fire = class Fire {
     dialog.getBuilder().get_object('settings-widget').connect('realize', (widget) => {
       const presets = [
         {
-          name: 'Default Fire',
+          name: _('Default Fire'),
           scale: 1.0,
           speed: 0.5,
           color1: 'rgba(76, 51, 25, 0.0)',
@@ -135,7 +137,7 @@ var Fire = class Fire {
           color5: 'rgba(255, 255, 255, 1)'
         },
         {
-          name: 'Hell Fire',
+          name: _('Hell Fire'),
           scale: 1.5,
           speed: 0.2,
           color1: 'rgba(0,0,0,0)',
@@ -145,7 +147,7 @@ var Fire = class Fire {
           color5: 'rgba(255, 255, 255, 1)'
         },
         {
-          name: 'Dark and Smutty',
+          name: _('Dark and Smutty'),
           scale: 1.0,
           speed: 0.5,
           color1: 'rgba(0,0,0,0)',
@@ -155,7 +157,7 @@ var Fire = class Fire {
           color5: 'rgb(255,238,166)'
         },
         {
-          name: 'Cold Breeze',
+          name: _('Cold Breeze'),
           scale: 1.5,
           speed: -0.1,
           color1: 'rgba(0,110,255,0)',
@@ -165,7 +167,7 @@ var Fire = class Fire {
           color5: 'rgb(97,189,255)'
         },
         {
-          name: 'Santa is Coming',
+          name: _('Santa is Coming'),
           scale: 0.4,
           speed: -0.5,
           color1: 'rgba(0,110,255,0)',